Turn `MediaList` into a factory-like reusable component.
