+++ src/components/RoomHistory/HistoryList.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import MediaList from '../MediaList';
+import withProps from 'recompose/withProps';
+import Base from '../MediaList/BaseMediaList';
 import AddToPlaylistAction from '../MediaList/Actions/AddToPlaylist';
 import HistoryRow from './Row';
 
@@ -14,14 +15,14 @@ const addMediaActions = onOpenAddMediaMenu =>
 
 const noActions = () => [];
 
-const HistoryList = ({ onOpenAddMediaMenu, ...props }) => (
-  <MediaList
-    {...props}
-    className="RoomHistory-list"
-    rowComponent={HistoryRow}
-    makeActions={onOpenAddMediaMenu ? addMediaActions(onOpenAddMediaMenu) : noActions}
-  />
-);
+const HistoryList = withProps(props => ({
+  className: 'RoomHistory-list',
+  listComponent: 'div',
+  rowComponent: HistoryRow,
+  makeActions: props.onOpenAddMediaMenu
+    ? addMediaActions(props.onOpenAddMediaMenu)
+    : noActions
+}))(Base);
 
 HistoryList.propTypes = {
   onOpenAddMediaMenu: PropTypes.func.isRequired,
